+++ src/view/edit_legislation.py
@@ -1,6 +1,10 @@
-from ..decorators import *
-from ..forms import *
+import logging
+from ..decorators import log_function_call
+from ..forms import LegislationForm
+from ..models import Legislation, ParliamentUser
 from django.shortcuts import render, get_object_or_404, redirect
+
+logger = logging.getLogger(__name__)
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponseForbidden