+++ src/view/upload_legislation.py
@@ -5,8 +5,8 @@ from django.core.exceptions import ValidationError
 from django.shortcuts import render, redirect
 from django.utils.dateparse import parse_datetime
 
-from ..decorators import *
-from ..forms import *
+from ..decorators import officer_required, log_function_call
+from ..forms import LegislationForm
 from src.models import Legislation, Role, ParliamentUser
 from src.utils.file_validation import validate_uploaded_file
 from src.notification_service import notify_all_active_members